ajout du réseau de neurone XOR

+++ Main/XOR/feedforward.c
@@ -1,3 +1,4 @@
+#include <stddef.h>
 #include "feedforward.h"
 #include "sigmoid.h"
 #include "matrix.h"
@@ -8,7 +9,7 @@ void feedforward(Network *net, double input[1][net->inputsize], double output[1]
 	matrix_product(1, net->inputsize, net->hiddensize, input, net->w1, net->values);
 	matrix_sum(1, net->hiddensize, net->b1, net->values, net->values);
 
-	int j;
+	size_t j;
 	for (j = 0; j < net->hiddensize; j++)
 		net->values[0][j] = sigmoid(net->values[0][j]);
